requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
pytest==7.4.4
python-dotenv==1.0.0
alembic